hal: audio_extn: Disable FWK enhancements based on property
Some devices may have FWK enhancements for some components, like
the IMS or video, but not for audio: in this case, sound will be
completely KO.

At this point, just add a property that default-disables the FWK
enhancements in the audio HAL, unless it's set true.

To enable FWK detection here, set "vendor.audio.full.fwk.enhance"
to true.

Change-Id: Ica878fdbcbd07e0b9a41d13ad190e77030667f24