hal: platform_info: Fall back to Pixel audio_platform_info if error
On some platforms, the audio_platform_info.xml file gets renamed
in many very creative ways.
If someone wants to stay clean, that someone is being forced into
finding what kind of creative names the HAL wants for the specific
sound card combination on the device, which is tricky and really
not an immediate nor the best way to do it: after all, if you are
bringing up audio for a project, you're not pushing configurations
that don't belong to that specific project, so why the... :)))

For this reason, introduce a fallback mechanism so that if the
Qualcomm creative suffix to "audio_platform_info" is not valid or
the developer hasn't played the find-the-name game, try to read
the standard "audio_platform_info.xml" file, as per Google's
guidelines.

Co-authored-by: Marijn Suijten <marijns95@gmail.com>
Change-Id: Ib25a34f3f0cffc230bae8eee93247216d4eb87e3